Implement the grab function.
This exercise left me profoundly confused.  I don't understand Coq
dependent pattern matching yet.  For example, I'm not sure why I need
to explicitly match on the first argument of UCons and why Coq is
happy to accept a single pattern (S n') but is not able to refined the
first argument automatically to be (S n) based on the type of ls.

I also tried to include the case for Nil, which should be impossible,
and use (False_rec (sig P) _) as the right hand side, but Coq
complained that it wasn't able to infer the placeholder _.  The error
went away when I completely removed the Nil branch.

I was also unable to figure out how to apply the convoy pattern from
the book here.